ml_project/requirements.txt
# local package
-e .

# external requirements
setuptools==50.3.2
numpy==1.19.4
pandas==1.1.3
scipy==1.5.2
sklearn==0.0
scikit-learn==0.24.1
PyYAML==5.3.1
dataclasses==0.6
omegaconf==2.0.6
pytest==6.1.2